+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# v0.3.0
+
+### Breaking Changes
+
+* Removed simulator file back into cyton repo
+* Removed `safe-buffer` in place of `buffer`
+* Removed support for node version 5 and below
+
+# v0.2.7
+
+### Bug Fixes
+
+* No `stopByte` property for daisy samples. Added tests.
+
+# v0.2.6
+
+### Bug Fixes
+
+* Accel data with counts did not work for cyton with daisy over wifi.
+
 # v0.2.5
 
 ### Bug Fixes
@@ -81,7 +101,7 @@
 
 ### New Functions
 
-* Add function `getFirmware(dataBuffer)` to utilities 
+* Add function `getFirmware(dataBuffer)` to utilities
 
 ### Breaking Changes
 
@@ -119,7 +139,7 @@
 ### New Features
 
 * In openBCIUtilities.js add function `transformRawDataPacketToSample` to parse a single raw data packet
-* In openBCIConstants.js add function `rawDataToSampleObjectDefault(numChannels)` which should be used by drivers to create the object that is passed through each call to `transformRawDataPacketsToSample` 
+* In openBCIConstants.js add function `rawDataToSampleObjectDefault(numChannels)` which should be used by drivers to create the object that is passed through each call to `transformRawDataPacketsToSample`
 
 # v0.0.6